requirements.txt
# requirements.txt

fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
pydantic<2.0.0
python-dotenv
docker
requests
